+++ commitlint.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  extends: ['gitmoji'],
-  rules: {
-    'subject-case': [2, 'always', 'sentence-case'],
-    'header-max-length': [2, 'always', 72]
-  }
-}