+++ packages/cache/src/defineWalletSetup.ts
@@ -1,4 +1,5 @@
 import type { BrowserContext, Page } from 'playwright-core'
+import buildWalletSetupFunction from './utils/buildWalletSetupFunction'
 import { getWalletSetupFuncHash } from './utils/getWalletSetupFuncHash'
 
 // TODO: Should we export this type in the `release` package?
@@ -15,7 +16,9 @@ export type WalletSetupFunction = (context: BrowserContext, walletPage: Page) =>
  * @returns An object containing the hash of the function, the function itself, and the wallet password. The `testWithWalletSetup` function uses this object.
  */
 export function defineWalletSetup(walletPassword: string, fn: WalletSetupFunction) {
-  const hash = getWalletSetupFuncHash(fn)
+  const walletSetupFunction = buildWalletSetupFunction(fn.toString())
+
+  const hash = getWalletSetupFuncHash(walletSetupFunction)
 
   return {
     hash,