+++ packages/cache/src/utils/buildWalletSetupFunction.ts
@@ -1,0 +1,15 @@
+import { transformSync } from 'esbuild'
+
+export default function buildWalletSetupFunction(walletSetupFunctionString: string) {
+  const { code } = transformSync(walletSetupFunctionString, {
+    format: 'esm',
+    minifyWhitespace: true,
+    target: 'es2022',
+    drop: ['console', 'debugger'],
+    loader: 'ts',
+    logLevel: 'silent',
+    platform: 'node'
+  })
+
+  return code
+}