+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @synthetixio/synpress-cache
 
+## 0.0.7
+
+### Patch Changes
+
+- Release
+
+## 0.0.6
+
+### Patch Changes
+
+- Release
+
+## 0.0.5
+
+### Patch Changes
+
+- Release
+
 ## 0.0.4
 
 ### Patch Changes