+++ wallets/ethereum-wallet-mock/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @synthetixio/ethereum-wallet-mock
 
+## 0.0.7
+
+### Patch Changes
+
+- Release
+- Updated dependencies
+  - @synthetixio/synpress-core@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- Release
+- Updated dependencies
+  - @synthetixio/synpress-core@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- Release
+- Updated dependencies
+  - @synthetixio/synpress-core@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes